requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pydantic==2.9.0
pydantic_core==2.23.2
